# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vacation Plans API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/holiday-plans": {
            "get": {
                "tags": [
                    "Holiday plans"
                ],
                "summary": "Get all plans",
                "operationId": "889175e6820800a1a254505c5f4f3049",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Halloween"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "costume party"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Wallstreet, 25"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "example": "YYYY-MM-DD"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Holiday plans"
                ],
                "summary": "Create a new plan",
                "operationId": "ac4e82aa2ced0698c793099bc45d8d92",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Halloween"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "costume party"
                                            },
                                            "location": {
                                                "type": "string",
                                                "example": "Wallstreet, 25"
                                            },
                                            "date": {
                                                "type": "string",
                                                "example": "YYYY-MM-DD"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Halloween"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "costume party"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Wallstreet, 25"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "example": "YYYY-MM-DD"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/holiday-plans/{id}": {
            "get": {
                "tags": [
                    "Holiday plans"
                ],
                "summary": "Get plan by ID",
                "operationId": "ba22c89a057eafb90a0e2cd63a5697b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Halloween"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "costume party"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Wallstreet, 25"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "example": "YYYY-MM-DD"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Holiday plans"
                ],
                "summary": "Edit a plan",
                "operationId": "1da83b8ebcb3026821d1cfcf01b78ce4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plan ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Halloween"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "costume party"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Wallstreet, 25"
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "YYYY-MM-DD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User edited successfully"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Holiday plans"
                ],
                "summary": "Delete plan by ID",
                "operationId": "a4949b7ccddf71caf457f0fd2e5baeb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plan ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Holiday plan deleted with success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/holiday-plans/get-pdf/{id}": {
            "get": {
                "tags": [
                    "Holiday plans"
                ],
                "summary": "Generate a pdf of the plan based on plan ID",
                "operationId": "2221d64895a09258c396ea7620a805d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plan ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generate a pdf in another page"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}